chore: refine code for kbcli related-config subcommand
